+++ src/valuta/exceptions.py
@@ -1,7 +1,10 @@
 __author__ = "Artur Barseghyan"
 __copyright__ = "2021 Artur Barseghyan"
 __license__ = "GPL-2.0-only OR LGPL-2.1-or-later"
-__all__ = ("ImproperlyConfigured",)
+__all__ = (
+    "ImproperlyConfigured",
+    "InvalidCurrency",
+)
 
 
 class ImproperlyConfigured(Exception):
@@ -9,3 +12,7 @@ class ImproperlyConfigured(Exception):
 
     Supposed to be thrown when code is improperly configured.
     """
+
+
+class InvalidCurrency(ValueError):
+    """Invalid currency exception."""