+++ src/valuta/tests/test_cli.py
@@ -1,4 +1,3 @@
-import logging
 import subprocess
 import tempfile
 import unittest
@@ -12,8 +11,6 @@ __author__ = "Artur Barseghyan"
 __copyright__ = "2021 Artur Barseghyan"
 __license__ = "GPL-2.0-only OR LGPL-2.1-or-later"
 __all__ = ("TestCLI",)
-
-LOGGER = logging.getLogger(__name__)
 
 
 class TestCLI(unittest.TestCase):