+++ src/valuta/tests/test_utils.py
@@ -1,4 +1,3 @@
-import logging
 import unittest
 
 from ..utils import get_currency_choices, get_currency_choices_with_code
@@ -13,8 +12,6 @@ __author__ = "Artur Barseghyan"
 __copyright__ = "2021 Artur Barseghyan"
 __license__ = "GPL-2.0-only OR LGPL-2.1-or-later"
 __all__ = ("TestUtils",)
-
-LOGGER = logging.getLogger(__name__)
 
 
 class TestUtils(unittest.TestCase):